Final Project - Initial 4

+++ FinalProject/FinalProject/Engine.swift
@@ -18,7 +18,7 @@ protocol EngineProtocol {
     var cols: Int { get set }
     var grid: GridProtocol { get }
     weak var delegate: EngineDelegate? { get set }
-    var refreshRate:  Double { get set }
+    var refreshRate:  Float { get set }
     var refreshTimer: NSTimer? { get set }
     func step() -> GridProtocol
     
@@ -58,7 +58,7 @@ class StandardEngine: EngineProtocol {
     
     weak var delegate: EngineDelegate?
     
-    var refreshRate: Double = 0.0
+    var refreshRate: Float = 0.0
     var refreshTimer: NSTimer?
     
     subscript (i: Int, j: Int) -> CellState {
